playground-server: refactored as controllers (except error handler)
